AG-17809 Move presets into folders
