+++ packages/ag-charts-enterprise/src/preset/gauge/gaugePreset.ts
