+++ packages/ag-charts-enterprise/src/preset/price-volume/__image_snapshots__/price-volume-preset-test-ts-price-volume-preset-create-financial-chart-for-candlestick-it-should-render-to-canvas-as-expected-1-snap.png
