+++ packages/ag-charts-enterprise/src/preset/price-volume/__image_snapshots__/price-volume-preset-test-ts-price-volume-preset-create-financial-chart-for-hollow-candlestick-it-should-render-to-canvas-as-expected-1-snap.png
