+++ packages/ag-charts-enterprise/src/preset/price-volume/__image_snapshots__/price-volume-preset-test-ts-price-volume-preset-create-financial-chart-for-minimal-it-should-export-image-data-url-as-expected-ag-12985-1-snap.png
